Extract test for enslig to its own class.
